Rename discovery project to feedResources
