+++ .github/workflows/feedResources.yml
@@ -7,7 +7,7 @@ concurrency:
 on:
   pull_request:
     branches: [ "develop" ]
-    paths: [ "discovery/**" ]
+    paths: [ "feedResources/**" ]
 
 jobs:
 
@@ -27,6 +27,6 @@ jobs:
 
       - name: Build with Gradle
         run: |
-          cd discovery
+          cd feedResources
           chmod +x ./gradlew
           ./gradlew build
