--- discovery/.idea/detekt.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DetektPluginSettings">
-    <option name="configurationFilePaths">
-      <set>
-        <option value="$PROJECT_DIR$/config/detekt/detekt-config.yml" />
-      </set>
-    </option>
-    <option name="enableAllRules" value="true" />
-    <option name="enableFormatting" value="true" />
-    <option name="treatAsErrors" value="true" />
-  </component>
-</project>
