+++ feedResources/app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <application
-        android:name=".ExampleApplication"
+        android:name="tv.fw.example.ExampleApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
@@ -11,7 +11,7 @@
         android:theme="@style/Theme.FireworkExample"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="tv.fw.example.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
