+++ feedResources/app/src/main/res/layout/details.xml
