+++ feedResources/app/src/main/res/values/strings.xml
