+++ feedResources/build.gradle.kts
