+++ feedResources/config/detekt/detekt-config.yml
