+++ feedResources/gradle.properties
