+++ feedResources/settings.gradle.kts
